Enabled docker system prune by default and enabled an overridabled switch

Related work items: #693190

+++ build/build-runtimeimages.sh
@@ -95,5 +95,8 @@ echo "List of images built (from '$RUNTIME_IMAGES_ARTIFACTS_FILE'):"
 cat $RUNTIME_IMAGES_ARTIFACTS_FILE
 
 echo
-echo "Cleanup: Running 'docker system prune' ..."
-docker system prune -f
+echo "Cleanup: Run 'docker system prune': $DOCKER_SYSTEM_PRUNE"
+if [ $DOCKER_SYSTEM_PRUNE = "true" ]
+then
+	docker system prune -f
+fi
